requirements.txt
python-dotenv
weaviate-client
pdfplumber

fastapi
uvicorn
dropbox
requests

PyMuPDF
python-multipart